[fix] 修复TbConst的parent字段错误地定义为uint16_t导致const表较大时溢出的bug。改成uint32_t。

+++ huatuo/metadata/Tables.h
@@ -148,7 +148,7 @@ namespace metadata
     struct TbConstant
     {
         uint8_t type; // 实际上占2字节
-        uint16_t parent;
+        uint32_t parent;
         uint32_t value;
     };
